requirements.txt
flask
requests
pypdf
ebooklib
python-docx
